feat(adapter-d1): split env and http drivers into separate files
