+++ packages/adapter-d1/src/constants.ts
@@ -1,0 +1,5 @@
+export const MAX_BIND_VALUES = 98
+
+// The SQLITE_ERROR result code is a generic error code that is used when no other more specific error code is available.
+// See: https://www.sqlite.org/rescode.html.
+export const GENERIC_SQLITE_ERROR = 1
